split test interface into several tests
