+++ test/test_tweets.rb
@@ -1,0 +1,21 @@
+require 'faker'
+
+module NanoTwitter
+  module Test
+    module Tweets
+
+      def self.registered(app)
+
+        app.get '/test/tweets/:num' do |num|
+          user = User.where(name: 'testuser').first
+          num.to_i.times do |i|
+            Tweet.create(user_id: user.id,
+                         content: Faker::Hacker.say_something_smart,
+                         tweeted_at: Faker::Time.backward(30, :all))
+          end
+          redirect '/'
+        end
+      end
+    end
+  end
+end
